Move view related types to dedicated module
